gym_hpa_deepsets/requirements.txt
numpy>=1.23.0
matplotlib>=3.5.2
pandas>=1.3.5
gymnasium
PettingZoo>=1.18.1
SuperSuit>=3.4.0
torch>=1.11.0
setuptools>=62.6.0
kubernetes==23.6.0
wheel==0.37.1
requests>=2.28.0
statsmodels>=0.13.2
scikit-learn>=1.1.1
keras>=2.9.0
joblib>=1.1.0
stable_baselines3>=1.5.0
tensorflow>=2.9.1


#Q: I uploaded this project to another computer and it said Illegal Instruction
